Update fork to upstream Dify 1.17.1

+++ web/app/components/base/prompt-editor/plugins/query-block/index.tsx
@@ -1,31 +1,13 @@
-import {
-  memo,
-  useEffect,
-} from 'react'
-import {
-  $insertNodes,
-  COMMAND_PRIORITY_EDITOR,
-  createCommand,
-} from 'lexical'
-import { mergeRegister } from '@lexical/utils'
-import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
 import type { QueryBlockType } from '../../types'
-import {
-  $createQueryBlockNode,
-  QueryBlockNode,
-} from './node'
+import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
+import { mergeRegister } from '@lexical/utils'
+import { $insertNodes, COMMAND_PRIORITY_EDITOR, createCommand } from 'lexical'
+import { memo, useEffect } from 'react'
+import { $createQueryBlockNode, QueryBlockNode } from './node'
 
 export const INSERT_QUERY_BLOCK_COMMAND = createCommand('INSERT_QUERY_BLOCK_COMMAND')
 export const DELETE_QUERY_BLOCK_COMMAND = createCommand('DELETE_QUERY_BLOCK_COMMAND')
-
-export type QueryBlockProps = {
-  onInsert?: () => void
-  onDelete?: () => void
-}
-const QueryBlock = memo(({
-  onInsert,
-  onDelete,
-}: QueryBlockType) => {
+const QueryBlock = memo(({ onInsert, onDelete }: QueryBlockType) => {
   const [editor] = useLexicalComposerContext()
 
   useEffect(() => {
@@ -39,8 +21,7 @@ const QueryBlock = memo(({
           const contextBlockNode = $createQueryBlockNode()
 
           $insertNodes([contextBlockNode])
-          if (onInsert)
-            onInsert()
+          if (onInsert) onInsert()
 
           return true
         },
@@ -49,8 +30,7 @@ const QueryBlock = memo(({
       editor.registerCommand(
         DELETE_QUERY_BLOCK_COMMAND,
         () => {
-          if (onDelete)
-            onDelete()
+          if (onDelete) onDelete()
 
           return true
         },
